Add support for mounting volumes concurrently
Mounting multiple GMUserFileSystem instances concurrently has not been
supported officially. It appeared to work, but it resulted in undefined
behavior when the file system process is killed.

Version 4.9.0 introduced changes (signal handling) that made it
impossible to mount multiple GMUserFileSystem instances at the same
time. This commit works around this limitation and addresses the
undefined behavior.