Avoid releasing objects we do not own
We used to over-retain the options dictionary passed to the mount
thread just so we can release it on that thread later. However, this
approach is considered poor style, especially since NSThread already
retains objects passed to it.

Starting with commit 92c8c06, the options dictionary is no longer being
over-retained, but it was still being released in the mount thread.
This mismatch led to a crash when unmounting a volume.

For details, see macfuse/macfuse#1086.